main task complete, working on optional hard task

+++ run_tests.sh
@@ -1,3 +1,5 @@
-rm test
+if [ -e "test" ]; then
+    rm test
+fi
 gcc -o test test.c palindrome.c /usr/local/src/unity.c -I/.
 ./test
